+++ changelog.md
@@ -4,6 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.8] - 2023-07-12
+
+- Enhancement: Add `svn:tag` command to create a new SVN tag for a plugin.
+
+## [1.2.7] - 2022-03-22
+
+- Enhancement: Create extensions as `tec-labs-` rather than `tribe-ext-`
+
 ## [1.2.6] - 2021-06-03
 ### Updated